Extract assertVisualMatch to its own file
